+++ nevergrad/optimization/base.py
@@ -3,13 +3,13 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-import abc
+import uuid
 import time
 import inspect
 import warnings
 from numbers import Real
 from collections import deque
-from typing import Optional, Tuple, Callable, Any, Dict, List, Union, Deque, Type
+from typing import Optional, Tuple, Callable, Any, Dict, List, Union, Deque, Type, Set
 import numpy as np
 from ..common.typetools import ArrayLike, JobLike, ExecutorLike
 from .. import instrumentation as instru
@@ -40,6 +40,7 @@ class Candidate:
         self.args = args
         self.kwargs = kwargs
         self.data = np.array(data, copy=False)
+        self.uuid = uuid.uuid4().hex
         self._meta: Dict[str, Any] = {}
 
     def __getitem__(self, ind: int) -> None:
@@ -117,11 +118,11 @@ class CandidateMaker:
         return Candidate(args, kwargs, data)
 
 
-class Optimizer(abc.ABC):  # pylint: disable=too-many-instance-attributes
+class Optimizer:  # pylint: disable=too-many-instance-attributes
     """Algorithm framework with 3 main functions:
-    - "ask()" which provides points on which to evaluate the function to optimize
-    - "tell(x, value)" which lets you provide the values associated to points
-    - "provide_recommendation()" which provides the best final value
+    - "ask()" which provides a candidate on which to evaluate the function to optimize
+    - "tell(candidate, value)" which lets you provide the values associated to points
+    - "provide_recommendation()" which provides the best final candidate
     Typically, one would call "ask()" num_workers times, evaluate the
     function on these num_workers points in parallel, update with the fitness value when the
     evaluations is finished, and iterate until the budget is over. At the very end,
@@ -168,8 +169,10 @@ class Optimizer(abc.ABC):  # pylint: disable=too-many-instance-attributes
         self.pruning: Optional[Callable[[utils.Archive[utils.Value]], utils.Archive[utils.Value]]] = None
         self.pruning = utils.Pruning.sensible_default(num_workers=num_workers, dimension=self.instrumentation.dimension)
         # instance state
+        self._asked: Set[str] = set()
         self._num_ask = 0
         self._num_tell = 0
+        self._num_tell_not_asked = 0
         self._callbacks: Dict[str, List[Any]] = {}
         # to make optimize function stoppable halway through
         self._running_jobs: List[Tuple[Candidate, JobLike[float]]] = []
@@ -186,6 +189,10 @@ class Optimizer(abc.ABC):  # pylint: disable=too-many-instance-attributes
     @property
     def num_tell(self) -> int:
         return self._num_tell
+
+    @property
+    def num_tell_not_asked(self) -> int:
+        return self._num_tell_not_asked
 
     def __repr__(self) -> str:
         inststr = f'{self.instrumentation:short}'
@@ -210,8 +217,8 @@ class Optimizer(abc.ABC):  # pylint: disable=too-many-instance-attributes
         """
         self._callbacks = {}
 
-    def tell_not_asked(self, x: Candidate, value: float) -> None:
-        """Provides the optimizer with the evaluation of a fitness value at a point it did not ask
+    def tell(self, candidate: Candidate, value: float) -> None:
+        """Provides the optimizer with the evaluation of a fitness value for a candidate.
 
         Parameters
         ----------
@@ -219,29 +226,26 @@ class Optimizer(abc.ABC):  # pylint: disable=too-many-instance-attributes
             point where the function was evaluated
         value: float
             value of the function
-        """
-        # default to just a tell
-        # algorithms which do not support it should raise NotImplementedError
-        self.tell(x, value)
 
-    def tell(self, args: Candidate, value: float) -> None:
-        """Provides the optimizer with the evaluation of a fitness value at a point
-
-        Parameters
-        ----------
-        x: np.ndarray
-            point where the function was evaluated
-        value: float
-            value of the function
+        Note
+        ----
+        The candidate should generally be one provided by ask(), but can be also
+        a non-asked candidate. To create a Candidate instance from args and kwargs,
+        you can use optimizer.create_candidate.from_call(*args, **kwargs)
         """
-        if not isinstance(args, Candidate):
-            raise TypeError("'tell' must be provided with the candidate that 'ask' provided")
-        x = args.data
+        if not isinstance(candidate, Candidate):
+            raise TypeError("'tell' must be provided with the candidate (use optimizer.create_candidate.from_call(*args, **kwargs)) "
+                            "if you want to inoculate a point that as not been asked for")
         # call callbacks for logging etc...
         for callback in self._callbacks.get("tell", []):
-            callback(self, x, value)
-        self._update_archive_and_bests(x, value)
-        self._internal_tell(x, value)
+            callback(self, candidate, value)
+        self._update_archive_and_bests(candidate.data, value)
+        if candidate.uuid in self._asked:
+            self._internal_tell_candidate(candidate, value)
+            self._asked.remove(candidate.uuid)
+        else:
+            self._internal_tell_not_asked(candidate, value)
+            self._num_tell_not_asked += 1
         self._num_tell += 1
 
     def _update_archive_and_bests(self, x: ArrayLike, value: float) -> None:
@@ -271,32 +275,64 @@ class Optimizer(abc.ABC):  # pylint: disable=too-many-instance-attributes
     def ask(self) -> Candidate:
         """Provides a point to explore.
         This function can be called multiple times to explore several points in parallel
+
+        Returns
+        -------
+        Candidate:
+            The candidate to try on the objective function. Candidates have field "args" and "kwargs" which can be directly used
+            on the function (objective_function(*candidate.args, **candidate.kwargs)).
         """
         # call callbacks for logging etc...
         for callback in self._callbacks.get("ask", []):
             callback(self)
-        suggestion = self._internal_ask()
-        assert suggestion is not None, f"{self.__class__.__name__}._internal_ask method returned None instead of a point."
+        candidate = self._internal_ask_candidate()
+        assert candidate is not None, f"{self.__class__.__name__}._internal_ask method returned None instead of a point."
         self._num_ask += 1
-        return self.create_candidate.from_data(suggestion)
+        self._asked.add(candidate.uuid)
+        return candidate
 
     def provide_recommendation(self) -> Candidate:
         """Provides the best point to use as a minimum, given the budget that was used
+
+        Returns
+        -------
+        Candidate
+            The candidate with minimal value. Candidates have field "args" and "kwargs" which can be directly used
+            on the function (objective_function(*candidate.args, **candidate.kwargs)).
         """
         return self.recommend()  # duplicate method
 
     def recommend(self) -> Candidate:
-        """Provides the best point to use as a minimum, given the budget that was used
+        """Provides the best candidate to use as a minimum, given the budget that was used.
+
+        Returns
+        -------
+        Candidate
+            The candidate with minimal value. Candidates have field "args" and "kwargs" which can be directly used
+            on the function (objective_function(*candidate.args, **candidate.kwargs)).
         """
         return self.create_candidate.from_data(self._internal_provide_recommendation(), deterministic=True)
+
+    def _internal_tell_not_asked(self, candidate: Candidate, value: float) -> None:
+        """Called whenever calling "tell" on a candidate that was not "asked".
+        Defaults to the standard tell pipeline.
+        """
+        self._internal_tell_candidate(candidate, value)
+
+    def _internal_tell_candidate(self, candidate: Candidate, value: float) -> None:
+        """Called whenever calling "tell" on a candidate that was "asked".
+        """
+        self._internal_tell(candidate.data, value)
+
+    def _internal_ask_candidate(self) -> Candidate:
+        return self.create_candidate.from_data(self._internal_ask())
 
     # Internal methods which can be overloaded (or must be, in the case of _internal_ask)
     def _internal_tell(self, x: ArrayLike, value: float) -> None:
         pass
 
-    @abc.abstractmethod
     def _internal_ask(self) -> ArrayLike:
-        raise NotImplementedError("Optimizer undefined.")
+        raise RuntimeError("Not implemented, should not be called.")
 
     def _internal_provide_recommendation(self) -> ArrayLike:
         return self.current_bests["pessimistic"].x
@@ -324,6 +360,12 @@ class Optimizer(abc.ABC):  # pylint: disable=too-many-instance-attributes
             print information about the optimization (0: None, 1: fitness values, 2: fitness values and recommendation)
         callback: callable
             callable called on the optimizer (self) at the end of each iteration (for user specific logging, etc)
+
+        Returns
+        -------
+        Candidate
+            The candidate with minimal value. Candidates have field "args" and "kwargs" which can be directly used
+            on the function (objective_function(*candidate.args, **candidate.kwargs)).
 
         Note
         ----