+++ nevergrad/optimization/recaster.py
@@ -225,7 +225,7 @@ class RecastOptimizer(base.Optimizer):
             raise RuntimeError(f"No message for evaluated point {x}: {self._messaging_thread.messages}")
         messages[0].result = value  # post the value, and the thread will deal with it
 
-    def tell_not_asked(self, x: base.Candidate, value: float) -> None:
+    def _internal_tell_not_asked(self, candidate: base.Candidate, value: float) -> None:
         raise base.TellNotAskedNotSupportedError
 
     def _internal_provide_recommendation(self) -> base.ArrayLike: